Add ostree_repo_regenerate_summary_with_options
This is an extensible version of ostree_repo_regenerate_summary_with_options().

Initial new options is just no-deltas-in-summary, which can override the
similar named config option.

Additionally there is also a callback for generating metadata. Both
for global metadata (entire summary file) and for each individual ref
in a ref table in the summary file.

Using a callback for the metadata allows clients to add per-commit metadata,
which wasn't initially possible. However, it will also later allow us to add
new versions of the summary file, and use different metadata in the
different version. For example, flatpak intends to change the way metadata
is stored in new summaries, but need to use the old format for the old
backwards compatible summary file.