summary: (optional) shorter key for commit timestamp
Rather than using "ostree.commit.timestamp" as the key name for the
commit timestamp in the metadata dict for every ref in the summary
we use the shorter "ot.cts".

The alignment for the variant after the key in the dic is 8, which
means the two options are 24 vs 8 bytes of size in the file. In
something like the flathub repo (with 10k refs) this ends up saving
80k, which is not neglible for something so simple to fix.

Again, this is off by default, but as I plan to change the summary
format we can easily enable this by default in the new summary files
(and off for the backwords compat one).