summary: Support by-ref-matched subsets
This lets you add subset to the repo config. We use a group name
similar to how the remote groups look like `[subset "foo"]`.

In the group, if there is a `refs` key it will be parsed as a regexp,
and only the matching refs will be part of the subset.

Here is an example subset config for a flatpak repo:
```
[subset "i386"]
refs=^((app|runtime)\/[.\\-_a-zA-Z0-9]+\/i386\/[.\\-_a-zA-Z0-9]+|appstream.?\/i386)$
```