summaries: Support subsets by matching a regexp on ref name
This lets you add subset to the repo config, named similar to the remote
groups. In the group, if there is a `refs` key it will be parsed as
a regexp, and only the matching refs will be part of the subset.

Here is an example subset config for a flatpak repo:
```
[subset "i386"]
refs=^((app|runtime)\/[.\\-_a-zA-Z0-9]+\/i386\/[.\\-_a-zA-Z0-9]+|appstream.?\/i386)$
```