Support indexed summaries wherever we fetch summaries
This allows the use of indexed summaries in:
  ostree_repo_remote_list_refs(),
  ostree_repo_remote_list_collection_refs(),
  ostree_avahi_service_build_repo_finder_result()
  ostree_repo_find_remotes_async()

These conversions are mostly trivial as there are no difference in the
formats as used. The only tricky part is in
ostree_repo_find_remotes_async() which needs a bit of help to get the
last_modified timestamp as that moved from the summary to the index.

It is important that we use the indexed summary everywhere if available
so we don't cache multiple summary formats.

The only internal remaining user of the old summary is the pull code.