summary: Add internal fetch_with_summary helper that also reports last_modified
The way this is detected differs in the indexed and non-indexed case. In
the non-indexed it is a metadata in the summary itself, whereas in the
indexed case this is in the index. We only need this for the p2p code
so we export this as an extra out parameter in an internal helper.