pull: Add old signature workaround support to fetch_non_indexed_summary()
Older versions of ostree could save the cached file before verifying
the signature, which lead to an unrecoverable state. The workaround
for this is in the pull code only, but this copies it to the
fetch_non_indexed_summary implementation too, in preparation for sharing
that code with the pull code.