summary: Always return and cache the index signature
Even if we are not currently verifying the signature of the index
repo_remote_fetch_summary_index() now returns it if found, and stores
it in the cache.

We want this because (as per e.g. test-pull-mirror-summary.sh) we want
to be able to mirror an summary that has signatures even if we're not
verifying atm.

In the verifying case nothing changes. However, it does mean that if
the gpg-verify-summary state changes from FALSE to TRUE we might have
cached an old version of the summary without validating its
signature. And at that point we will continue to trust it until the
upstream summary index changes and we will then verify the new one. I
think this is fair behaviour, as you did at some point trust that the
non-verified summary was ok.

We never run into a progress problem with stale invalid signatures
(like we used to in c4c2b5eb), because we never re-verify the cached
index.