summary: Generate indexed summaries when updating
By default we now generate the old backwards compatible summary, as
well as a new indexed and content addressable
OSTREE_SUMMARY_VERSION_INDEXED format with files in the "summaries"
subdirectory. The `max-summary-format-version` repo option can disable
generation of the new summary format.

The format for the index (`summary.idx`) is a map from a subset name
to information about the partial summary for that subset. The per summary
information is the checksum of the summary file (accessible as
`summaries/$(sha256).summary`) and an array of previous summary checksums
(useful for deltas). There is also a traditional a{sv} additional
metadata field in the file.

Right now the only generated subset is `""` which contains everything,
but the long term plan is to allow subsets (such as "all foss apps" or
"only ARM apps").

The summary file format is identical to the previous summaries except they
always enable the no-deltas-in-summary and new_commit_timestamp_key feature.
Also, the `ostree.summary.last-modified` field was moved from the summary
into the index, allowing the individual summaries to be reproducible if
nothing changes in a particular subset inbetween updates.

The `metadata_callback` callback from ostree_repo_regenerate_summary_with_options()
will give a version 0 when updating the old summary file, and 1 when updating the
new ones. This will let e.g. flatpak generate different formats for the metadata
in the new and old summaries.

During summary update we delete old summaries that are not refered from the newly
generated as well as the on disk one. This means any outstanding operations for
the previous summary will continue working while still evantually cleaning
up old summaries.

The code that signs the summary now also signs the `summary.idx` file
if it is there. Additionally, since the APIs for summaries really want
to return a summary for the actual index file we sign those too. This
is also used by flatpak where it passes the signed summary file to the
privileged system-helper.

The `summary.idx` is updated after all individual summaries are
written, so in terms of atomicity it is of similar status to the old
summary file. I.e. care must be takes to not cache it forever, and ensure it
is not cached out of sync with `summary.idx.sig`. However the individual
content-addressed summaries are immutable and have no caching issues.