pull: Support using indexed summaries when pulling
This uses the pre-existing helpers for fetching and caching summary
indexes their corresponding summaries.

Some special handling:
 * We disable caches when pulling from a local repo
 * In the mirror case we download both the indexed and regular summary
 * We need to return the index and the summary checksum from the index
   lookup so that we know where save the indexed summary when mirroring.

There is also a `disable-indexed-summaries` option in the pull API which
we're going to need for the tests.

ostree pull now supports --disable-indexed-summaries which we need
for the tests.

The tests are updated so that they test both the indexed and non-indexed
summaries, and look for the caches etc in the right place for each.